feat: separate zoom into a shared util file (#606)
* feat: separate zoom into a shared util file

* refactor: more changes

* refactor: updating config
