+++ projects/observability/src/shared/components/topology/d3/d3-topology.ts
@@ -181,7 +181,8 @@ export class D3Topology implements Topology {
       container: svg,
       target: data,
       scroll: this.config.zoomable ? zoomScrollConfig : undefined,
-      pan: this.config.zoomable ? zoomPanConfig : undefined
+      pan: this.config.zoomable ? zoomPanConfig : undefined,
+      showBrush: true
     });
 
     this.onDestroy(() => {
@@ -281,7 +282,7 @@ export class D3Topology implements Topology {
     topologyData.nodes.forEach(node => nodeRenderer.drawNode(groupElement, node));
     topologyData.edges.forEach(edge => edgeRenderer.drawEdge(groupElement, edge));
     topologyData.nodes.forEach(node => this.select(nodeRenderer.getElementForNode(node)!).raise());
-    this.zoom.updateBrushOverlay(topologyData.nodes);
+    this.zoom.updateBrushOverlayWithData(topologyData.nodes);
   }
 
   private updateMeasuredDimensions(): void {
